Fix a bug where lint-staged would not run type checking for a frontend application
Since the frontend application contains a tsconfig.json file with an empty array as the value of
the "files" key, type checking does not make sense when running tsc with this file as a
configuration file. The solution is to increase the number of files that are considered
configuration files and run type checking with all such files as configuration files.