refactor(backend): change session deletion logic
Since the session is used to determine the last time the user was online, it is necessary to store
at least 1 session for each user. So when the user somehow requests to delete a session that is the
last one, reset the max age of that session instead of deleting it.